drivers: dma: siwx91x: Fix DMA_ADDR_ADJ_NO_CHANGE w/ memory
GPDMA does not support DMA_ADDR_ADJ_NO_CHANGE with a memory buffer. This
feature is required fro the SPI driver. Hopefully, SPI driver is the only
user of this feature.

So, this commit introduces a hack for SPI. Rather than copying memory, it
configure the DMA to fill the destination memory (with either 0s or 1s).
  - For MISO: the destination buffer is ignored by the driver
  - For MOSI, we only support support values 0x00 or 0xFF. Hopefully, none
    will need ay other value.

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>